module_2: TrelloApi class
